move tf override tensor to torch_function.py
