+++ torch/_dynamo/variables/torch.py
@@ -41,8 +41,11 @@ from .ctx_manager import (
 from .distributed import is_constant_pg_functions, is_from_local, ProcessGroupVariable
 from .higher_order_ops import TorchHigherOrderOperatorVariable
 from .lists import ListVariable, TupleVariable
-from .tensor import TensorWithTFOverrideVariable
-from .torch_function import can_dispatch_torch_function, dispatch_torch_function
+from .torch_function import (
+    can_dispatch_torch_function,
+    dispatch_torch_function,
+    TensorWithTFOverrideVariable,
+)
 
 log = logging.getLogger(__name__)
